Working Export Command with not OOP design
