+++ src/main/java/seedu/address/model/event/DateTime.java
@@ -20,7 +20,7 @@ public class DateTime {
                     + "Date: 1 - 31\n"
                     + "Hour: 0-23\n"
                     + "Minute: 0-59\n";
-    public static final DateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy HH:mm");
+    private static final DateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy HH:mm");
 
     public final Date dateTime;
